Add javascript to stop minuteCountDown() instance that has been updated by the user. Before the old and new instances would compete to change the innerHTML --On child_added an instance attribute with a value of 1 is given to each train row. MinuteCountdown() checks the instance value at start and saves the value to a variable, then compares the saved value to the actual DOM value every minute. On child_changed the instance attribute value is increased by 1 at every change. If minuteCountDown() compares a value greater that the initial, it stops the instance so the newer instance can take over.
Add checkCurrentTime() to make sure that train start time is present day time && before current time. --On child_added the train start time is sent to checkCurrentTime to validate the start time. If train start time is a time after the current time, it is assumed the user is talking about yesterday and so checkCurrentTime() takes the train start time and forwards it to present day and feeds that start time to calculateTrainArrivalTime(). If time is not validated the train arrival time comes back incorrect.